Big ol' session refactor.  We no longer have separate records for sessions and users.
git-svn-id: file:///export/repos/witgap/trunk@233 be230a1c-449f-4488-b524-6273046f66f0

+++ src/cpp/server/db/ActorRepository.h
@@ -21,7 +21,7 @@ public:
      * Initializes the repository, performing any necessary migrations.
      */
     void init ();
-    
+
     /**
      * Finds actors whose names start with the specified prefix.  The callback will receive a
      * ResourceDescriptorList containing the actor descriptors.
@@ -29,7 +29,7 @@ public:
      * @param creatorId the id of the creator whose actors are desired, or 0 for all creators.
      */
     Q_INVOKABLE void findActors (
-        const QString& prefix, quint32 creatorId, const Callback& callback);
+        const QString& prefix, quint64 creatorId, const Callback& callback);
 };
 
 #endif // ACTOR_REPOSITORY
